requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-cloud-bigquery==3.14.1
python-dotenv==1.0.1
pandas==2.2.0